Move integration tests into a separate package
Jira: MODKBEKBJ-344
