+++ src/test/java/org/folio/rest/impl/IntegrationTestSuite.java
@@ -7,6 +7,21 @@ import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
+import org.folio.rest.impl.integrationsuite.DefaultLoadHoldingsImplTest;
+import org.folio.rest.impl.integrationsuite.DefaultLoadServiceFacadeTest;
+import org.folio.rest.impl.integrationsuite.EHoldingsProxyTypesImplTest;
+import org.folio.rest.impl.integrationsuite.EHoldingsRootProxyImplTest;
+import org.folio.rest.impl.integrationsuite.EholdingsCacheImplTest;
+import org.folio.rest.impl.integrationsuite.EholdingsConfigurationTest;
+import org.folio.rest.impl.integrationsuite.EholdingsCustomLabelsImplTest;
+import org.folio.rest.impl.integrationsuite.EholdingsPackagesTest;
+import org.folio.rest.impl.integrationsuite.EholdingsProvidersImplTest;
+import org.folio.rest.impl.integrationsuite.EholdingsResourcesImplTest;
+import org.folio.rest.impl.integrationsuite.EholdingsStatusTest;
+import org.folio.rest.impl.integrationsuite.EholdingsTagsImplTest;
+import org.folio.rest.impl.integrationsuite.EholdingsTitlesTest;
+import org.folio.rest.impl.integrationsuite.LoadHoldingsStatusImplTest;
+import org.folio.rest.impl.integrationsuite.TransactionLoadServiceFacadeTest;
 import org.folio.test.util.TestSetUpHelper;
 
 @Suite.SuiteClasses({
@@ -24,7 +39,6 @@ import org.folio.test.util.TestSetUpHelper;
   DefaultLoadHoldingsImplTest.class,
   LoadHoldingsStatusImplTest.class,
   DefaultLoadServiceFacadeTest.class,
-  TransactionLoadHoldingsImplTest.class,
   TransactionLoadServiceFacadeTest.class
 })
 @RunWith(Suite.class)
