+++ src/test/java/org/folio/rest/impl/integrationsuite/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains tests that are run as part of IntegrationTestSuite.
+ * By default test classes in this package are ignored during build.
+ * To make sure that tests run during build add them to org.folio.rest.impl.IntegrationTestSuite.
+ */
+package org.folio.rest.impl.integrationsuite;
